MAILBOX-355 Move SystemMailboxesProvider to Mailbox-Api
